TASK 4 by ProdigyInfoTech
In this task of sentiment analysis, I fetch the data from API, add a new column of sentiment and sentiment_numeric, then visualize the data with the help of histogram, line graph and word cloud. I furtherly cleaned the data by adding steps of text preprocessing and lastly I trained and test the data to for the models of Logistic Regression, Random Forest and Naive Bayes.